Added a couple of functions to work with the mysql API, but my local computer didn't have the necessary library (mysql/mysql.h) so I was not able to run the code. In the end I just added a simple function that asks the user some basic information (server, user, password, and database), but did nothing with this info. git log

+++ general.h
@@ -9,7 +9,9 @@
 #define general_h
 
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
+//#include <mysql/mysql.h>
 
 /* Structure definition */
 
@@ -20,7 +22,7 @@ typedef struct user{
 
 /* Function prototypes */
 
-int verify_date_format(char date[11]);
-int verify_int_format(int to_verify, int min, int max);
+size_t verify_date_format(char date[11]);
+size_t verify_int_format(int to_verify, int min, int max);
 
 #endif /* general_h */
